+++ tests/UtilsTest.php
@@ -49,20 +49,47 @@ class UtilsTest extends TestCase
         self::assertSame('foobaz', (string) $s2);
     }
 
-    public function testStopsCopyToStreamWhenWriteFails(): void
+    public function testCopyToStreamRetriesShortWrites(): void
     {
         $s1 = Psr7\Utils::streamFor('foobaz');
-        $s2 = Psr7\Utils::streamFor('');
-        $s2 = FnStream::decorate($s2, [
-            'write' => function () {
-                return 0;
+        $sink = Psr7\Utils::streamFor('');
+        $writes = 0;
+
+        $s2 = FnStream::decorate($sink, [
+            'write' => function (string $string) use ($sink, &$writes): int {
+                ++$writes;
+
+                return $sink->write(substr($string, 0, 1));
             },
         ]);
+
         Psr7\Utils::copyToStream($s1, $s2);
-        self::assertSame('', (string) $s2);
+
+        self::assertSame('foobaz', (string) $sink);
+        self::assertSame(6, $writes);
     }
 
-    public function testStopsCopyToSteamWhenWriteFailsWithMaxLen(): void
+    public function testCopyToStreamRetriesShortWritesWithMaxLen(): void
+    {
+        $s1 = Psr7\Utils::streamFor('foobaz');
+        $sink = Psr7\Utils::streamFor('');
+        $writes = 0;
+
+        $s2 = FnStream::decorate($sink, [
+            'write' => function (string $string) use ($sink, &$writes): int {
+                ++$writes;
+
+                return $sink->write(substr($string, 0, 1));
+            },
+        ]);
+
+        Psr7\Utils::copyToStream($s1, $s2, 3);
+
+        self::assertSame('foo', (string) $sink);
+        self::assertSame(3, $writes);
+    }
+
+    public function testCopyToStreamThrowsWhenWriteFails(): void
     {
         $s1 = Psr7\Utils::streamFor('foobaz');
         $s2 = Psr7\Utils::streamFor('');
@@ -71,8 +98,27 @@ class UtilsTest extends TestCase
                 return 0;
             },
         ]);
+
+        $this->expectException(\RuntimeException::class);
+        $this->expectExceptionMessage('Unable to write to stream');
+
+        Psr7\Utils::copyToStream($s1, $s2);
+    }
+
+    public function testCopyToStreamThrowsWhenWriteFailsWithMaxLen(): void
+    {
+        $s1 = Psr7\Utils::streamFor('foobaz');
+        $s2 = Psr7\Utils::streamFor('');
+        $s2 = FnStream::decorate($s2, [
+            'write' => function () {
+                return 0;
+            },
+        ]);
+
+        $this->expectException(\RuntimeException::class);
+        $this->expectExceptionMessage('Unable to write to stream');
+
         Psr7\Utils::copyToStream($s1, $s2, 10);
-        self::assertSame('', (string) $s2);
     }
 
     public function testCopyToStreamReadsInChunksInsteadOfAllInMemory(): void